Add test_identity
