+++ tests/conftest.py
@@ -9,6 +9,28 @@ from torch.utils.data import DataLoader, Dataset
 from auto_cast.types import Batch, EncodedBatch
 
 
+def _make_batch(
+    batch_size: int = 1,
+    t: int = 1,
+    w: int = 2,
+    h: int = 3,
+    c: int = 2,
+    const_c: int = 1,
+    scalar_c: int = 1,
+) -> Batch:
+    input_fields = torch.arange(batch_size * t * w * h * c, dtype=torch.float32)
+    input_fields = input_fields.view(batch_size, t, w, h, c)
+    output_fields = torch.zeros(batch_size, t, w, h, c)
+    constant_fields = torch.ones(batch_size, w, h, const_c)
+    constant_scalars = torch.full((batch_size, scalar_c), 5.0)
+    return Batch(
+        input_fields=input_fields,
+        output_fields=output_fields,
+        constant_scalars=constant_scalars,
+        constant_fields=constant_fields,
+    )
+
+
 def assert_output_valid(output: Tensor, expected_shape: tuple, name: str = "Output"):
     """Assert output has expected shape and contains no NaN values."""
     assert output.shape == expected_shape, (
